requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cx_Oracle==8.3.0
fastapi==0.112.1
h11==0.14.0
idna==3.7
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.38.2
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
